authorization: move it towards Authorized{Keys,Principals}Command
Move all the authorization code to both AuthorizedKeysCommand  pubkey
based authentication and AuthorizedPrincipalsCommand for cert based
authentication.

The pam_oslogin_admin.so module is removed as it's not required any
more and pam_oslogin_login.so is now only used for 2fa authentication.

This change requires a configuration addaptation to guest-agent as it now
requires adding AuthorizedPrincipalsCommand{User} and dropping where
it was adding pam_oslogin_admin.so to /etc/pam.d/sshd as well as only
adding pam_oslogin_login.so for 2fa auth.